fix linting errors
