+++ frontend/src/components/portraits/ItemPortrait.vue
